LayoutTests/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/close/close-basic-expected.txt

CONSOLE MESSAGE: line 23: WebSocket connection to 'ws://localhost:49001/' failed: WebSocket is closed before the connection is established.

FAIL WebSockets: close() assert_equals: expected 2 but got 3
